pyproject.toml
[tool.poetry]
name = "python-practical-guide"
version = "0.1.0"
description = ""
authors = ["Michal Slota <michal.slota@elevup.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pycryptodome = "^3.14.1"
Flask = "^2.1.1"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
